+++ frontend/src/api/getPortfolio.js
@@ -1,5 +1,5 @@
-import { setPortfolio } from "./portfolioDetailSlice";
-import ApplicationContext from "../../../applicationContext/ApplicationContext";
+import { setPortfolio } from "../store/portfolioDetailSlice";
+import ApplicationContext from "../applicationContext/ApplicationContext";
 
 export const getPortfolio = async (id) => {
     const api_version = ApplicationContext.get().helpers().backEndConfig.apiVersion;
