+++ frontend/src/api/getPortfolio.test.js
@@ -1,7 +1,7 @@
 import { getPortfolioAndSetState } from "./getPortfolio";
-import store from "../../../store";
-import TestApplicationContext from "../../../applicationContext/TestApplicationContext";
-import { dispatchUsecase } from "../../../helpers/test";
+import store from "../store";
+import TestApplicationContext from "../applicationContext/TestApplicationContext";
+import { dispatchUsecase } from "../helpers/test";
 
 test("successfully gets the Portfolio from the backend and directly puts it into state", async () => {
     const mockPortfolioId = "2";
