+++ frontend/src/index.jsx
@@ -10,7 +10,7 @@ import "./uswds/css/styles.css";
 
 import Home from "./pages/Home";
 import PortfolioList from "./pages/portfolios/list/PortfolioList";
-import PortfolioDetail from "./pages/portfolios/detail/PortfolioDetail";
+import PortfolioDetail from "./pages/PortfolioDetail";
 import CanList from "./pages/cans/list/CanList";
 import CanDetail from "./pages/cans/detail/CanDetail";
