--- frontend/src/pages/portfolios/detail/styles.module.css
@@ -1,4 +1,0 @@
-body {
-    width: 1024px;
-    margin: auto;
-}
