+++ frontend/src/store.js
@@ -3,7 +3,7 @@ import canListSlice from "./pages/cans/list/canListSlice";
 import canDetailSlice from "./pages/cans/detail/canDetailSlice";
 import canFiscalYearSlice from "./pages/cans/detail/budgetSummary/canFiscalYearSlice";
 import portfolioListSlice from "./pages/portfolios/list/portfolioListSlice";
-import portfolioDetailSlice from "./pages/portfolios/detail/portfolioDetailSlice";
+import portfolioDetailSlice from "./store/portfolioDetailSlice";
 import portfolioFundingSummarySlice from "./components/PortfolioFundingSummary/portfolioFundingSummarySlice";
 // eslint-disable-next-line import/no-named-as-default
 import authSlice from "./components/Auth/authSlice";
